singleFileInfo.js
